fix: refactor structure
